requirements.txt
fastapi
uvicorn[standard]

langchain
langchain-core
langchain-community
langchain-groq
langchain-text-splitters
langchain-huggingface

faiss-cpu
sentence-transformers
huggingface-hub

ragas
datasets

python-dotenv
pandas
scikit-learn

streamlit